Create matfree/funm.py and matfree/funm_trace.py to collect matrix-function functionality
