+++ tests/test_funm/test_funm_lanczos_spd.py
@@ -1,10 +1,10 @@
 """Test matrix-function-vector products via Lanczos' algorithm."""
 
-from matfree import lanczos, test_util
+from matfree import funm, test_util
 from matfree.backend import linalg, np, prng
 
 
-def test_funm_vector_product(n=11):
+def test_funm_lanczos_spd(n=11):
     """Test matrix-function-vector products via Lanczos' algorithm."""
     # Create a test-problem: matvec, matrix function,
     # vector, and parameters (a matrix).
@@ -12,8 +12,6 @@ def test_funm_vector_product(n=11):
     def matvec(x, p):
         return p @ x
 
-    # todo: write a test for matfun=np.inv,
-    #  because this application seems to be brittle
     def fun(x):
         return np.sin(x)
 
@@ -29,6 +27,6 @@ def test_funm_vector_product(n=11):
 
     # Compute the matrix-function vector product
     order = 6
-    matfun_vec = lanczos.funm_vector_product_spd(fun, order, matvec)
+    matfun_vec = funm.funm_lanczos_spd(fun, order, matvec)
     received = matfun_vec(v, matrix)
     assert np.allclose(expected, received, atol=1e-6)
