+++ tests/test_funm_trace/test_integrand_logdet_product.py
@@ -1,6 +1,6 @@
 """Test stochastic Lanczos quadrature for log-determinants of matrix-products."""
 
-from matfree import hutchinson, lanczos, test_util
+from matfree import funm_trace, hutchinson, test_util
 from matfree.backend import linalg, np, prng, testing
 
 
@@ -31,7 +31,7 @@ def test_logdet_product(A, order):
 
     x_like = {"fx": np.ones((ncols,), dtype=float)}
     fun = hutchinson.sampler_normal(x_like, num=400)
-    problem = lanczos.integrand_product_logdet(order, matvec, vecmat)
+    problem = funm_trace.integrand_product_logdet(order, matvec, vecmat)
     estimate = hutchinson.hutchinson(problem, fun)
     received = estimate(key)
 
@@ -53,7 +53,7 @@ def test_logdet_product_exact_for_full_order_lanczos(n):
 
     # Set up max-order Lanczos approximation inside SLQ for the matrix-logarithm
     order = n - 1
-    integrand = lanczos.integrand_product_logdet(
+    integrand = funm_trace.integrand_product_logdet(
         order, lambda v: A @ v, lambda v: v @ A
     )
